Modularize anchor logic
