+++ app/components/sidebar/anchor_menu_component.html.erb
@@ -1,0 +1,21 @@
+<li class="nav-item nav-item-has-children" data-sidebar-target="groupList" data-controller="sidebar-anchor-menu">
+  <a
+    href="#0"
+    class="<%= @class %>"
+    data-bs-toggle="collapse"
+    data-bs-target="#ddmenu_<%= @identifier %>"
+    aria-controls="ddmenu_<%= @identifier %>"
+    aria-expanded="true"
+    aria-label="Toggle navigation"
+    data-sidebar-anchor-menu-target="title">
+              <span class="icon">
+                <i class="lni mr-10 lni-<%= @icon %>"></i>
+              </span>
+    <span data-sidebar-target="linkTitle"><%= @title %></span>
+  </a>
+  <ul id="ddmenu_<%= @identifier %>" class="collapse dropdown-nav" data-sidebar-anchor-menu-target="list">
+    <% links.each do |link| %>
+      <%= link %>
+    <% end %>
+  </ul>
+</li>
